lab_final/app/requirements.txt

fastapi==0.95.0
uvicorn==0.21.1
google-cloud-pubsub==2.15.0
pydantic==1.10.7
